Now uses Steam for usernames
Game attempts to use steam to get username, if it cannot get to it for some reason, it uses the name last stored in PlayerPrefs, which is normally the steam name.

The player can still change their name manually.

Added 'Pack Folder' button, which brings the player to their pack folder.